+++ client/components/ForgotPassword.js
@@ -1,77 +1,102 @@
-import React, {Component, Fragment} from 'react'
-import PropTypes from 'prop-types'
+import React from 'react'
+import {connect} from 'react-redux'
+import {postForgotPassword} from '../store'
 
-import Avatar from '@material-ui/core/Avatar'
 import Button from '@material-ui/core/Button'
-import CssBaseline from '@material-ui/core/CssBaseline'
+import Dialog from '@material-ui/core/Dialog'
+import DialogActions from '@material-ui/core/DialogActions'
+import DialogContent from '@material-ui/core/DialogContent'
+import DialogContentText from '@material-ui/core/DialogContentText'
+import DialogTitle from '@material-ui/core/DialogTitle'
+import {withStyles} from '@material-ui/core/styles'
 import FormControl from '@material-ui/core/FormControl'
 import Input from '@material-ui/core/Input'
-import InputLabel from '@material-ui/core/InputLabel'
-import LockIcon from '@material-ui/icons/LockOutlined'
-import Paper from '@material-ui/core/Paper'
 import Typography from '@material-ui/core/Typography'
-import {withStyles} from '@material-ui/core/styles'
 
 const styles = theme => ({
-  paper: {
-    // marginTop: theme.spacing.unit * 8,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    padding: `${theme.spacing.unit * 2}px ${theme.spacing.unit * 3}px ${theme
-      .spacing.unit * 3}px`
-  },
   submit: {
     marginTop: theme.spacing.unit * 3
-  },
-  avatar: {
-    margin: theme.spacing.unit,
-    backgroundColor: theme.palette.secondary.main
-  },
-  form: {
-    width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing.unit
   }
 })
 
-class ForgotPassword extends Component {
+class ForgotPassword extends React.Component {
+  state = {
+    open: false,
+    email: ''
+  }
+
+  handleClickOpen = () => {
+    this.setState({open: true})
+  }
+
+  handleChange = event => {
+    this.setState({[event.target.name]: event.target.value})
+  }
+  handleClose = () => {
+    this.setState({open: false})
+  }
+
   handleSubmit = event => {
     event.preventDefault()
+
+    this.props.postForgotPassword({email: this.state.email})
   }
   render() {
-    const {classes} = this.props
+    const {classes, status} = this.props
+
     return (
-      <Fragment>
-        <CssBaseline />
-        <Paper className={classes.paper}>
-          <Avatar className={classes.avatar}>
-            <LockIcon />
-          </Avatar>
-          <Typography component="h1" variant="h5">
-            Forgot Password
-          </Typography>
-          <form
-            className={classes.form}
-            onSubmit={this.handleSubmit}
-            name={name}
-          >
-            <FormControl margin="normal" required fullWidth>
-              <InputLabel htmlFor="email">Email</InputLabel>
-              <Input id="email" name="email" autoComplete="email" autoFocus />
-            </FormControl>
-            <Button
-              type="submit"
-              variant="contained"
-              color="primary"
-              className={classes.submit}
-            >
-              SEND EMAIL
-            </Button>
+      <div>
+        <Button className={classes.submit} onClick={this.handleClickOpen}>
+          Forgot Password
+        </Button>
+        <Dialog
+          open={this.state.open}
+          onClose={this.handleClose}
+          aria-labelledby="form-dialog-title"
+        >
+          <form onSubmit={this.handleSubmit}>
+            <DialogTitle id="form-dialog-title">Forgot Password</DialogTitle>
+            <DialogContent>
+              <DialogContentText>
+                Please enter your email to reset your password.
+              </DialogContentText>
+              <FormControl>
+                <Input
+                  autoFocus
+                  margin="dense"
+                  id="email"
+                  name="email"
+                  placeholder="Email Address"
+                  onChange={this.handleChange}
+                  value={this.state.email}
+                  label="Email Address"
+                  type="email"
+                  fullWidth
+                />
+              </FormControl>
+              <Typography>{status && status.data && status.data}</Typography>
+            </DialogContent>
+            <DialogActions>
+              <Button onClick={this.handleClose} color="primary">
+                Close
+              </Button>
+              <Button variant="contained" type="submit" color="primary">
+                Submit
+              </Button>
+            </DialogActions>
           </form>
-        </Paper>
-      </Fragment>
+        </Dialog>
+      </div>
     )
   }
 }
 
-export default withStyles(styles)(ForgotPassword)
+const mapDispatchToProps = dispatch => ({
+  postForgotPassword: email => dispatch(postForgotPassword(email))
+})
+
+const mapStateToProps = state => ({status: state.password.status})
+
+export default connect(mapStateToProps, mapDispatchToProps)(
+  withStyles(styles)(ForgotPassword)
+)
