+++ client/components/auth-form.js
@@ -15,7 +15,7 @@ import LockIcon from '@material-ui/icons/LockOutlined'
 import Paper from '@material-ui/core/Paper'
 import Typography from '@material-ui/core/Typography'
 import {withStyles} from '@material-ui/core/styles'
-import ForgotPassword2 from './ForgotPassword2'
+import {ForgotPassword} from '../components'
 
 const styles = theme => ({
   main: {
@@ -138,7 +138,7 @@ const AuthForm = props => {
               >
                 Forgot Password
               </Button> */}
-            <ForgotPassword2 />
+            <ForgotPassword />
             <div />
 
             <Button
