+++ client/components/index.js
@@ -15,4 +15,3 @@ export {default as AddParty} from './AddParty'
 export {default as EditParty} from './EditParty'
 export {default as Gallery} from './Gallery'
 export {default as ForgotPassword} from './ForgotPassword'
-export {default as ForgotPassword2} from './ForgotPassword2'
